flush-response-body: avoid the need for multiple callbacks
In the case of a streaming body the code can decide to either
immediately flush the response line and headers, or wait until the user
explicitly signals a flush. The latter has better batching behavior, and
so is preferred, at least for now.

The previous implementation of this behavior incorrectly assumed that
multiple writer callbacks could be registered with response body. This
was true in the past but is no longer the case. To get around this, the
the code forces the `Writer.t`'s underlying `Faraday.t` to produce a
`Yield` as its next operation, and then invokes the callback that was
registered with the `Reqd.t` immediately. This will cause the conneciton
runtime to receive the yield, and register a new writer callback with
the connection. This new callback will make its way to the
Response.Body.t that was just created, and will be invoked when the user
calls `flush` on it.