parse-invariant: do buffer bookkeeping in report_read_result
It was unclear when and where the buffer read/write window was being
updated, which caused assertion errors in the code after the last
refactoring. This commit makes it clear that `report_read_result` is
responsbile for zeroing out the "committed" byte count in the parser
state, and shifting the read/write window of the input buffer
accordingly.