request-eof-fix: read handlers must consume all input
This is unforunate but currently necessary. In order to use the
underlying Faraday.t to buffer input, a bit more thought needs to be put
into the interface. For now, reads are one-shot. Either something must
be done with all the bytes while the on_read function has control, or
the have to be copied.