fix-iovec-shiftv: change alcotest naming convention
The naming convetion in this the test_httpaf module requires test
writers to think about the order in which tests appear in a submodule.
To avoid this, prefix the name of the funciton in the module that's
being tested with `test_`.

This came up while writing tests for the IOVec module, which will appear
in a subsequent commit.