fix-iovec-shiftv: add tests that demonstarate the bug
Pull request #33 purported to fix a bug, and did identify the presence
of a bug, but did not completely address it. It was not sufficient to
swap the comparison, but also modify the false branch to only shift off
`n` from the iovec at the head of the list, rather than the length of
the iovec itself.

The test also identifies a misleading exception message, namely

  "shiftv: n >= lengthv iovecs"

That should be a strict inequality.

Both these issues will be fixed in subsequent commits.