fix-chunked-encoding: fix the serializer to emit the final chunk
This fix is suboptimal in a few ways. First, it introduces another match
on the body encoding. Rather than checking the encoding every time, the
body should just cache this value, or have client code pass in just a
boolean indicating whether or not the body should be chunk-encoded.

Secondly, emitting the final chunk requires one additional write poll
regardless of whether the body is chunk-encoded. This should be avoided
for non-chunk encoded bodies at the very least, and ideally for
chunk-encoded bodies as well. But for now this works.