set `write_final_if_chunked` to `false` in Body if not Chunked encoding
Currently for persistent connections, we're always trying to poll one
more write even if the response body isn't chunked. This causes weird
issues such as #37.

This PR attempts to fix this problematic behavior. I wasn't able to produce
a failing test, most likely because the test simulator always runs synchronously,
but it does fix the issue for me locally.