fix double callback in when_ready_to_write
When you yield a writer, the callback is either stored in one of three
places, or it is invoked immediately by `Body.when_ready_to_write`. The
immediate invocation happens if the body is closed, but that conditional
does not prevent the rest of the function from running. In some cases,
you may immediately invoke a callback, store it in
`t.when_ready_to_write`, and then later invoke it again later when that
field is checked.

This also moves the check for `t.when_ready_to_write` before the
synchronous invocation. In the case where we have a pending callback, it
would be nice for the two possible outcomes of `when_ready_to_write` to
behave similarly.