add a test for flushing streaming bodies synchronously
This test exposes an issue where a response hangs indefinitely because
`respond_with_streaming` puts the writer into `yield` mode so that it
doesn't write until the first `flush`, but if the first `flush` call
happens before the write loop "exhausts" the yield, it will just get
stuck.

This requires a change in Faraday [1] to fix.

[1]: https://github.com/inhabitedtype/faraday/pull/44