fix advancing request queue
Due to a complicated chain of if/then/else expressions, the
`advance_request_queue_if_necessary` function was structured in such a
way that the branch to handle requests ending a persistent connection
was actually being applied to non-complete requests.

This fix exposed a bug in the tests where the `connection: close` header
would cause the request to be closed immediately. This was because
`Reqd.is_complete` is true as long as there is nothing to be read or
written -- that is, the request body is complete and the entire response
body has made it to `t.writer` but _not necessarily flushed_. Combine
this with the fact that the proceeding `shutdown` would close the
writer, causing it to be drained of pending writes, and you have
connections that close immediately.

We fix this by changing the behavior of shutdown. Now it only closes the
writer and does not drain it, which allows the external world to flush
it as normal. This is a nice fit, since as it is now, closing the reader
still allows the server to finish reading the request it is currently
working on.