Client bindings: don't shutdown_send immediately
This is related to #77 and #78, but independent enough to warrant a new
pull request.

When using http/af over HTTPS, the SSL / TLS library needs to keep the
socket open for writing because there might be writes on the upgraded
connection.

This changeset proposes that the http/af `Client` modules don't close
the socket for sending, and instead wait until the request is over to
shutdown the file descriptor.