remove-handle-type-param: approach has been abandoned
The handle parameter of a server connection was originally intended to
support upgrades within the HTTP state machine. This approach was never
actually tested for feasiblity. In the mean time a different approach of
building support for upgrading into the runtime has been tested for
feasiblity, and will be adopted instead.