Fix race condition when reporting errors, and `report_exn` in the Async bindings
Whenever `Server_connection.next_write_operation` is called before
`Server_connection.next_read_operation` in the Async / Lwt bindings and
the request handler throws an exception, the writer has yielded and
won't wake up to write the exception to the socket. This solution is to
wake up the writer explicitly.