upgrade: support upgrades via an I/O operation
Add `Reqd.respond_with_upgrade`, which will create response with a with
status code 101, and the provided headers. On the next write operation,
this response, along with the inspiring request will be provided to the
runtime. It is up to the runtime to serialize the resposne and send it
on the wire before handing off the socket to an upgrade handler.