test-response-parser: fix raise in response parser
The status-code parser consumed input until it reached a space. In the
event that a non-alphanumeric character appeared in the status-code, the
subsequent int conversion would raise. This also happens in the case
where the status-message is omitted, since the next character after the
status-code is a CR.

The status-code parser has been fixed to only consume digits, which
should make the int conversion safe.