input-shrunk-on-exn: demonstrate input shrunk when reporting client exn
When the parser has been fed input but has not hit a commit point so the
entire buffer can be consumed, a subseqent report_exn call will result
in the parser raising with the message "input shrunk".

This is a requirement of angstrom. After it sees some input, you can't
take it away.

Fixed in a subsequent commt.