add output state variant to reqd
Similar to recent changes that added `Input_state.t`, this adds an
`Output_state.t` with an additional `Waiting` state, as, unlike input
data, the output may be waiting for a request handler to decide how or
what to write to the body. Even though we changed `requires_output` to
return an output state instead, this function was no longer used in
`Server_connection` after the recent changes moving the write loop
continuation directly to the writer. So this change is fairly trivial,
but it is a necessary step in supporting upgrade requests as we do
additional refactoring of the server and its request queue handling
semantics.

This also seemed like a good opportunity to rename `Complete`, since it
is really a dual to `Streaming`, and the name felt a little misleading
(and in fact has tripped me up on at least one occasion). I thought that
`Fixed` was good, as you respond with a single fixed string or bigstring
which gets written directly to the writer. Perhaps `Direct` or something
in that vein would work as well (though that feels less like a "state"
and more like a "kind").

I also imported some minor style tidying from other pull requests since
I was touching these match statements already.