move yield to writer
Previously we did this dance where whenever the writer is in a yielded
state, we start by holding on to it in the server connection and then
passing it off to the reqd when it's created. And then *that* was
sometimes handed off to the `Body` so that it could call it when the
user invoked `Body.flush` or `Body.close_and_drain`. And then each
further yield, we have to figure out who should be holding on to it. It
gets a little confusing and forces us to do some extra invariant checking.

Now we put the yield callback in neither place and instead keep it on
the writer directly. There is exactly one writer per server connection,
and it is also the place we ask for `next` once the writer is awake. So
it seems like a much more natural place to keep it.

We stil have to consider the body, though. We simply keep the callback
mechanism around in the body, but when it would normally be invoking its
single-use thunk, it instead invokes a persistent thunk that `Reqd` set
up in order to wake the writer. This was something we didn't get right
the first pass, and we realized that all tests were passing incorrectly,
so we added a missing asynchronous streaming test.

We also added an assertion to `next` that it is never called while the
writer is yielded. This broke a few tests that use
`next_write_operation` as a means of probing the state to make sure it's
still yielded, so we cleaned those up to test a ref instead.

Since body was previously used by creating and then immediately
attaching a `when_ready_to_write` callback, we should merge that into
create. But since we use the body for both reading and writing, I didn't
want to take an unused argument when the body was creating for reading
purposes. So we now have separate functions for creating the body in the
reader or writer role, though I think we should go a step further and
make them separate types.

In doing this, I realized that it makes sense to make use of the
[Writer]'s [on_wakeup] callback in the client connection, too, which was
previously wiring those callbacks up to the body. So now the client and
server state machines behave similarly in that regard.