expose parsed headers in the order they were read
We recursively read off headers from requests (in the case of a server)
and responses (in the case of a client) and construct a list out of
them, but because `Headers.t` is an exposed type, we were able to pass
an ordered list when it should have been reversed. This change adds a
simple mli that matches the interface of `Httpaf.Headers`, which causes
a build error in one location in the parser where we need to reify the
headers from the alist we have. And since we use `lift2 cons`, we have
an ordered list. I also added a test that makes sure they are read in
the correct order.