add failing test for read errors
There are two kinds of read errors: parse errors (where angstrom
couldn't parse the request into a structured `Request.t`) and bad
requests, where it could parse the request but some invariant was not
satisfied, like having a positive body-length header if specifed.

It's not clear what the right behavior is in these cases, but currently
the writer will hang and never close.