branch-free-equal: add case-insensitive char equality check
Exhaustively check all character pairs and ensure that the naive
implementation of case-insensitive equality checking is consistent with
the one used in the Headers module. Getting a value for the latter is
accomplished by constructing a Headers.t that contains just one entry,
with its name being one of the characters to check, and checking for
membership using the other character.