ignore parse failures at checkpoints
Previously, transition would examine `state`, and if it represented a
parse failure with 0 consumed bytes, it would transition to `Done`. This
meant that if a request failed to parse right at a checkpoint, the
request would be silently dropped and the reader would be ready for a
new request. At this point the runtime probably fed the
previously-problematic line back into the reader for another error.

It's not clear why this branch exists. Back in 52bb5795 this was moved
from `next` to `transition`, where it previously asserted that the
parser was closed. I think that this was perhaps a way of dealing with
the reader encountering an EOF and not being able to complete the parse,
so closing the connection made sense. But this same operation doesn't
make sense in `transition`. We can remove the case entirely and let
another case handle it.

Add a test for the repro in #193. Without this patch, it fails at
`reader_closed t` because the reader is actually ready. If you re-feed
the same bad input, the request handler is called.