add test for responding mid-read
This was an untested case that was exposed in another PR. We should be
able to respond before reading the body. Add some test helpers for
headers and replace usages for consistency. I added a currently-unused
`writer_ready` assertion because it's nice when you're trying to probe
the state to debug.