Fix parse error at EOF.
Currently, if there is a parse error after [read_eof], the runtime has no
chance to discover it. The error is not surfaced immediately, and if
the runtime calls [next_read_operation], it just gets [`Close].
This is because [Reader.next] first checks to see if it is closed, and
if so, always returns [`Close].

This feature fixes this by inverting the logic in that function.

Unfortunately this leads to another problem: the parser effectively
expects an unending stream of requests, because after finishing
parsing one request it immediately expects the next one. This means
that read_eof is guaranteed to lead to a parse error. Thankfully this
is easily fixed by treating an empty input as a valid as well.