Don't hang the reader thread after write eof
In many runtimes, there are separate reader and writer threads that
drive the reading and writing from httpaf independently. So a thing
that can happen is:

- A request arrives.
- The response handler begins but does not finish responding to it.
- The writer thread calls [Server_connection.report_write_result
`Closed].
- The reader thread delivers another request.

In this case, httpaf will never deliver the second request to the
handler, because the first request's output_state never gets to
Complete. We have no hope of responding to the second request, but we
should still deliver it to the handler in case the request has
side-effects (e.g. as many POST requests do).

This PR fixes this by noticing when the writer is closed in
[Reqd.output_state] and just always returning [Complete] in this case,
as no more output progress can be made.