Changes to yielding behaviour to fix a test.
As a result of the above changes, the
[test_empty_fixed_streaming_response] test in
test_server_connection.ml broke. I think the fact it was working
before was just a pretty weird coincidence. It went like this:

1. When the Body_writer.t was created, [write_final_if_chunked] is set
to true.

2. The response body is created and the test's response handler is
called. This immediately closes the body.

3. The test runtime calls [next_write_operation], which checks
[Body_writer.has_pending_output]. This returns true -- which is wrong!
-- because at this point, [write_final_if_chunked] has not yet been
set to false.

4. At some point, [Body_writer.transfer_to_writer] gets called which
does now set [write_final_if_chunked] to false.

I'm not sure what to make of all of this. The test seems reasonable
enough -- if you close the body immediately, you should expect that
the output appears on the writer immediately.

I chose to fix this by removing calls to [Writer.yield] and
[Writer.unyield]. These seem kinda hard to think about: the writer
will be asked to yield until the next time the runtime calls
[next_write_operation]. I removed them and all the tests still passed,
so hopefully this is okay. However, I fully admit that I don't really
understand why they're there. If it's important for performance
somehow, then let me know and I can think of a different approach.