add `make watch` for development
I use this for fast feedback while developing and thought it would be
good to put in the `Makefile`. I would have built the examples and
benchmarks too, but all the linking makes the whole iteration take an
order of magnitude longer. Maybe there's some dune option I'm missing
here to let us build but not link.