body-writer-simplifications: simple reverts and tweaks
I reverted the change of `when_ready_to_write` to storing a reference to
the `writer`. I don't necessarily disagree with it, but it seemed like
an orthogonal change that made the other simplifications harder to
review. Plus I'm not entirely sure the separation of concerns there is
right, and there's a potential issue with using `of_faraday` and passing
the writer at the same time. This will be shown in another commit.

I also renamed `Fixed` to `Identity`, since that's the name of the
default transfer encoding -- the idea being that the encoding is how you
alter the body. Plus, `Close_connect -> Fixed` as a mapping felt
misleading.

I also brought back the `unyield` when the body transfer detects a
close. I think this is probably not used, but again I wanted to be more
conservative with other changes.