demonstrate error_handler encoding misuse
In the error handler, we create a body using `of_faraday`, meaning it's
directly wired to the request writer. The idea here was probably that if
there's no `reqd` (e.g., in the case of a parse error), there's no way
to know how to transfer the body.

But because of that, we never use `transfer_to_writer`, which means we
would basically treat a chunked encoding as close-delimited. These
responses are probably unreadable by clients, but I guess it's also rare
enough to use a chunked encoding in a response that it doesn't happen in
practice.

In another branch, I introduced an `is_errored` boolean to the server
connection, but that could potentially be changed to a variant that
references the body in the case of an error. Then we could transfer it
to the writer as with normal reqd response bodies.