improve error test coverage
I realized a recent branch that added the `is_errored` boolean actually
didn't have a test that showed when it is used. It came with a test, but
that test actually passed regardless. This is because there are two ways
to invoke `set_error_and_handle`: either the runtime uses `report_exn`,
or the parser enters an error state and the read loop discovers that.
The latter always means that `next_read_action` does *not* shut down the
writer, which is evident because `set_error_and_handle` is followed by a
`Close` return.

A test was added for the former. Along the way I thought of a few extra
scenarios to test, including one that might be new -- the
`error_handler` must be used to complete a request if invoked, or the
state machine will hang. I didn't port these onto the master from a week
ago to see how it behaves.

There were a few XXXs left in older error-handling tests from before the
request queue was refactored. I think that this is better understood now
and can just be summarized by saying the connection is shutdown by the
end of the test.