fix httpaf-async upgrade implementation
We don't want to invoke the handler separately in the read loop and
write loop, we want to wait until both are upgraded and then invoke.

I included a script that wraps `nc` as a way of testing. We don't have
upgrade support on the clients yet, and we can't use something like
`wscat` without also implementing the entire websocket stack, which is
more trouble than it seems worth right now.

This still has an issue that you can easily demonstrate with the
included script -- the "hello" message gets read into the `Buffer` along
with the headers and never passed along to the client. I don't know how
we should really address this, but we can't put data back onto the `fd`
that we intend to pass on to the handler, so we have to surface the
buffered data somehow.