move reader yields to separate function
This is purely so that we can attach the big comment that describes why
this is weird. The "improvements" that I mentioned in one of the
comments are lost to time, but I believe it involved allowing the reader
to be put into a yielded state once the parsing was complete and resumed
once more input was requested. Ultimately, we won't spend the time
trying to implement such a thing.